refactor: move composition stuffs to standalone package
