+++ app/src/main/java/com/osfans/trime/ime/composition/CompositionPopupWindow.kt
@@ -1,4 +1,4 @@
-package com.osfans.trime.ime.text
+package com.osfans.trime.ime.composition
 
 import android.content.Context
 import android.graphics.RectF
